Fix misleading log message for sync restart condition
The log message "Should restart sync" was misleading because it didn't
accurately reflect the condition it was checking. The condition
`lastSyncTimestamp < previous` (where `previous` is the previous value
of `lastSyncTimestamp`) is true not only when a sync needs to be
restarted due to a timestamp regression, but also when a sync needs to
be started for the first time (i.e., `previous` is null and
`lastSyncTimestamp` has an initial value).

This commit updates the log message to "Should start or restart sync" to
accurately reflect both scenarios.